+++ strader.h
@@ -1,0 +1,4 @@
+                                                         
+void run_target(const char* target);
+
+void run_debugger(int pid, int attach);
